vision_service/requirements.txt
fastapi==0.109.2
uvicorn==0.25.0
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
numpy==1.26.3
python-multipart==0.0.6
